[hail] Remove PackEncoder from the main package
It's only used for a few tests
